+++ plan/moving_bodies_test.go
@@ -61,7 +61,7 @@ func testContext(t *testing.T) *coord.Context {
 		t.Fatalf("NewGeodetic: %v", err)
 	}
 
-	site, err := NewSite("Test", loc, angle.Zero(), nil)
+	site, err := NewSite("Test", loc, nil)
 	if err != nil {
 		t.Fatalf("NewSite: %v", err)
 	}